requirements.txt
fastapi==0.99.1
openai==0.27.8
openai-whisper @ git+https://github.com/openai/whisper.git@b91c907694f96a3fb9da03d4bbdc83fbcd3a40a4
PyAudio==0.2.13
PyMySQL==1.1.0
uvicorn==0.22.0
Wave==0.0.2
whisper==1.1.10
pydub==0.25.1
PyMySQL==1.1.0
pyannote-whisper @ git+https://github.com/yinruiqing/pyannote-whisper.git
faster_whisper==0.7.1